requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.40.0
numpy>=1.26.0
tqdm>=4.66.0
pydantic>=2.7.0

python-json-logger
google-cloud-storage


# Data & Utils
pandas>=2.2.0
datasets>=2.19.0
scikit-learn>=1.4.0   # common utilities for metrics, preprocessing

# Logging & Config
rich>=13.7.0
pyyaml>=6.0.1
loguru>=0.7.2         # cleaner structured logging (optional but recommended)

# Testing
pytest>=8.2.0

# Optional (development tools)
jupyter>=1.0.0
black>=24.4.0
isort>=5.13.0
pre-commit>=3.7.0     # enforce linting/formatting before commits
